# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Presensi API",
        "description": "API documentation for the Presensi application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        },
        {
            "url": "/docs"
        }
    ],
    "paths": {
...
        "/yf/presensi": {
            "get": {
                "tags": [
                    "Presensi"
                ],
                "summary": "Load employee presence data",
                "description": "Memuat data absensi karyawan berdasarkan ID, bulan, dan tahun",
                "operationId": "687c3099216d622933b624ee4b99102f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID karyawan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Bulan dalam format 2 digit (default: bulan saat ini)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "01"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Tahun dalam format 4 digit (default: tahun saat ini)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2023"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil memuat data absensi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "employees_id": {
                                                        "type": "integer",
                                                        "example": 123
                                                    },
                                                    "presence_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2023-11-01"
                                                    },
                                                    "presence_status": {
                                                        "type": "string",
                                                        "example": "Present"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-11-01T12:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-11-01T12:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data absensi berhasil dimuat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data absensi tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data absensi tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}